Created an automated repo environment
This adds the ability for the user to stand up a repo for all of the python packages
that may be needed for a given environment. This will allow a user to maintain their
own repo of packages for production purposes in CDC where public internet may not be
available or for development purposes.

* The repo created by these plays will create a python wheel archive and a git server powered by NGINX enables support on port 8181 via haproxy for testing, and moves the wheel builder scripts into the ``pkg_repo`` roles.
* The plays also have support for multi node environments which will provide for scalability of the internal repo. This is provided by lsync on the multiple nodes.
* The repo storage is mounted in /openstack on the host.
* The cloudserver-aio script has been updated with the new container type.

Related Issue: https://github.com/rcbops/u-suk-dev/issues/48
Related Issue: https://github.com/rcbops/ansible-lxc-rpc/issues/527

Other updates impacted by the repo changes:
* udpated md5sum in environment file.
* added cron for report clean up weekly
* changed working directory and made memcache detect memory
* changed upstream url for repo
* Updated pip configuration layout in common
* incorporates changes made bt @nrb in commit d232585a062803c7033ab4dff482eec572dbd74updated memcached configs
* added documentation to the python script